[FEAT] 보관함 - 스와이프 완료
