+++ app/src/main/java/com/android/hara/presentation/home/fragment/storage/StorageFragment.kt
@@ -4,11 +4,14 @@ import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.View
+import androidx.viewpager2.widget.ViewPager2
 import com.android.hara.R
 import com.android.hara.databinding.FragmentStorageBinding
 import com.android.hara.presentation.base.BindingFragment
 import com.android.hara.presentation.home.fragment.storage.model.StorageData
 import com.android.hara.presentation.onesec.OneSecActivity
+import com.google.android.material.tabs.TabLayout
+import com.google.android.material.tabs.TabLayoutMediator
 
 
 class StorageFragment : BindingFragment<FragmentStorageBinding>(R.layout.fragment_storage) {
@@ -21,27 +24,25 @@ class StorageFragment : BindingFragment<FragmentStorageBinding>(R.layout.fragmen
     // sel에 최종적으로 selected가 전달되는 것
     // [결론] Fragment의 local = .xml의 selected = BindingConversion.kt의 sel
 
-    val ing = true // 고민중이면 true, 고민완료면 false //TODO 서버통신 시 변경
-
-    private val dummyList = listOf<StorageData>(
-        StorageData("일상", "우리집에 왜왔니 왜왔니 왜왔니", "2023.12.25", true),
-        StorageData("사랑", "카레는 맛있어 나는 안 졸려", "2023.12.25", true),
-        StorageData("일상", "우리집에 호이호이", "2023.12.25", true),
-        StorageData("시하", "우리집에 갤럭시", "2023.12.25", true),
-        StorageData("하시", "나나나나나나나나나나나마", "2023.12.25", false),
-        StorageData("사과", "울라울라울라울라울라울라울라울라", "2023.12.25", false),
-        StorageData("귤귤", "노트북노트북노트북노트북노트북노트북", "2023.12.25", false),
-        StorageData("안드", "노틀담의꼽추노틀담의꼽추", "2023.12.25", false)
-    )
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val storageAdapter = StorageAdapter() {
-            Log.d("TEST", it)
-        }
-        binding.rvPosts.adapter = storageAdapter
-        storageAdapter.submitList(dummyList)
+        // ViewPager2, TabLayout 참조
+        val viewPager: ViewPager2 = binding.viewpager
+        val tabLayout: TabLayout = binding.tlStorageTab
+
+        // FragmentStateAdapter 생성
+        val viewpagerFragmentAdapter = ViewpagerFragmentAdapter(requireActivity())
+
+        // ViewPager2의 adapter 설정
+        viewPager.adapter = viewpagerFragmentAdapter
+
+        // ###### TabLayout과 ViewPager2를 연결
+        // 1. 탭메뉴의 이름을 리스트로 생성해둔다.
+        val tabTitles = listOf<String>("혼자 고민", "함께 고민")
+
+        // 2. TabLayout과 ViewPager2를 연결하고, TabItem의 메뉴명을 설정한다.
+        TabLayoutMediator(tabLayout, viewPager, {tab, position -> tab.text = tabTitles[position]}).attach()
 
         // 버튼 클릭 시 '1초만에 해결하기' 액티비티로 연결
         binding.cvBtnOnesec.setOnClickListener {
