+++ app/src/main/java/com/android/hara/presentation/onesec/adapter/OneSecAdapter.kt
@@ -29,6 +29,4 @@ class OneSecAdapter() :
             binding.setVariable(BR.worrydata, getItem(position) as WorryData)
         }
     }
-
-
 }
